+++ console/packages/starwhale-core/src/datastore/hooks/useDatastoreTables.tsx
@@ -2,18 +2,6 @@ import React, { useEffect, useMemo } from 'react'
 import { useListDatastoreTables } from './useFetchDatastore'
 import { tablesOfEvaluation } from '../utils'
 
-export default function useDatastoreTables(projectName: string, jobUuid: string) {
-    const queryAllTables = useMemo(() => {
-        if (!projectName || !jobUuid) return {}
-        return {
-            prefix: tablesOfEvaluation(projectName, jobUuid),
-        }
-    }, [projectName, jobUuid])
-
-    // eslint-disable-next-line @typescript-eslint/no-use-before-define
-    return useDatastoreTablesByPrefix(queryAllTables?.prefix as string)
-}
-
 export function useDatastoreTablesByPrefix(prefix: string) {
     const allTables = useListDatastoreTables({ prefix })
