+++ console/packages/starwhale-ui/src/Search/filter.tsx
@@ -1,52 +1,11 @@
-import { StatefulMenu } from 'baseui/menu'
-import { PLACEMENT, Popover, StatefulPopover } from 'baseui/popover'
-import Select from '../Select'
-import { KIND, Operators } from './constants'
-import { FilterPropsT, FilterT } from './types'
 import { useState, useEffect } from 'react'
+import { StatefulMenu } from 'baseui/menu'
+import { PLACEMENT, Popover } from 'baseui/popover'
+import { FilterT } from './types'
+// eslint-disable-next-line import/no-cycle
 import { useStyles } from './Search'
-
-function Filter(options: FilterT): FilterT {
-    const operatorOptions = options.operators.map((key: string) => {
-        const operator = Operators[key]
-        return {
-            id: operator.key,
-            type: operator.key,
-            label: operator.label,
-        }
-    })
-
-    return {
-        kind: options.kind,
-        operators: options.operators,
-        renderField: function RenderField({ options = [], isEditing = false, ...rest }) {
-            return (
-                <PopoverContainer
-                    {...rest}
-                    options={options}
-                    isOpen={isEditing}
-                    onItemSelect={({ item }) => rest.onChange?.(item.type)}
-                >
-                    {rest.value}
-                </PopoverContainer>
-            )
-        },
-        renderOperator: function RenderOperator({ isEditing = false, ...rest }) {
-            return (
-                <PopoverContainer
-                    {...rest}
-                    options={operatorOptions}
-                    isOpen={isEditing}
-                    onItemSelect={({ item }) => rest.onChange?.(item.type)}
-                >
-                    {operatorOptions.find((v) => v.type === rest.value)?.label ?? ''}
-                </PopoverContainer>
-            )
-        },
-        // renderFieldValue: options?.renderFieldValue ?? <></>,
-    }
-}
-export default Filter
+// eslint-disable-next-line import/no-cycle
+import { Operators } from './utils'
 
 function PopoverContainer(props: {
     onItemSelect?: (props: { item: { label: string; type: string } }) => void
@@ -99,3 +58,45 @@ function PopoverContainer(props: {
         </Popover>
     )
 }
+
+function Filter(options: FilterT): FilterT {
+    const operatorOptions = options.operators.map((key: string) => {
+        const operator = Operators[key]
+        return {
+            id: operator.key,
+            type: operator.key,
+            label: operator.label,
+        }
+    })
+
+    return {
+        kind: options.kind,
+        operators: options.operators,
+        renderField: function RenderField({ options: renderOptions = [], isEditing = false, ...rest }) {
+            return (
+                <PopoverContainer
+                    {...rest}
+                    options={renderOptions}
+                    isOpen={isEditing}
+                    onItemSelect={({ item }) => rest.onChange?.(item.type)}
+                >
+                    {rest.value}
+                </PopoverContainer>
+            )
+        },
+        renderOperator: function RenderOperator({ isEditing = false, ...rest }) {
+            return (
+                <PopoverContainer
+                    {...rest}
+                    options={operatorOptions}
+                    isOpen={isEditing}
+                    onItemSelect={({ item }) => rest.onChange?.(item.type)}
+                >
+                    {operatorOptions.find((v) => v.type === rest.value)?.label ?? ''}
+                </PopoverContainer>
+            )
+        },
+        // renderFieldValue: options?.renderFieldValue ?? <></>,
+    }
+}
+export default Filter
