+++ console/packages/starwhale-ui/src/Search/filterNumberical.tsx
@@ -1,6 +1,5 @@
-import { Popover } from 'baseui/popover'
-import Input from '../Input'
 import { FilterTypeOperators, KIND } from './constants'
+// eslint-disable-next-line import/no-cycle
 import Filter from './filter'
 import { FilterT } from './types'
