+++ console/packages/starwhale-ui/src/Search/index.ts
@@ -1,1 +1,2 @@
+export { default as Search } from './Search'
 export { default } from './Search'
