+++ console/packages/starwhale-ui/src/base/select/autosize-input.tsx
@@ -4,6 +4,7 @@ Copyright (c) Uber Technologies, Inc.
 This source code is licensed under the MIT license found in the
 LICENSE file in the root directory of this source tree.
 */
+/* eslint-disable */
 
 import * as React from 'react'
 import { StyledInput, StyledInputSizer } from './styled-components'
